+++ src/main/kotlin/dev/willebrands/intellij/sloppyfocus/settings/SloppyFocusSettingsState.kt
@@ -1,0 +1,7 @@
+package dev.willebrands.intellij.sloppyfocus.settings
+
+import com.intellij.openapi.components.*
+
+class SloppyFocusSettingsState : BaseState() {
+    var focusDelayMs by property(200)
+}